# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ACR - Attack Chain Reconstruction Engine",
    "description": "Attack Chain Reconstruction Engine: ingest telemetry, run rule-based detection, correlate events with multi-signal scoring, reconstruct attack chains, map them to MITRE ATT&CK and evaluate against synthetic ground truth.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/chains": {
      "get": {
        "tags": [
          "chains"
        ],
        "summary": "List Chains",
        "operationId": "list_chains_api_chains_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "is_attack",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Attack"
            }
          },
          {
            "name": "scenario_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scenario Id"
            }
          },
          {
            "name": "host",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Host"
            }
          },
          {
            "name": "min_confidence",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 100,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Confidence"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChainListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/scenarios/generate": {
      "post": {
        "tags": [
          "scenarios"
        ],
        "summary": "Generate",
        "operationId": "generate_api_scenarios_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioGenerateOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/scenarios/reset": {
      "post": {
        "tags": [
          "scenarios"
        ],
        "summary": "Reset",
        "operationId": "reset_api_scenarios_reset_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioResetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioResetOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChainListOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ChainSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "ChainListOut"
      },
      "ChainSummary": {
        "properties": {
          "chain_id": {
            "type": "string",
            "title": "Chain Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "is_attack": {
            "type": "boolean",
            "title": "Is Attack",
            "default": false
          },
          "start_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time"
          },
          "end_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "event_count": {
            "type": "integer",
            "title": "Event Count",
            "default": 0
          },
          "confidence": {
            "$ref": "#/components/schemas/ConfidenceOut"
          },
          "risk": {
            "$ref": "#/components/schemas/RiskOut"
          },
          "tactics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tactics"
          },
          "techniques": {
            "items": {
              "$ref": "#/components/schemas/TechniqueRef"
            },
            "type": "array",
            "title": "Techniques"
          },
          "hosts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Hosts"
          },
          "users": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Users"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "scenario_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Id"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "chain_id",
          "status",
          "confidence",
          "risk"
        ],
        "title": "ChainSummary"
      },
      "ConfidenceOut": {
        "properties": {
          "score": {
            "type": "number",
            "title": "Score"
          },
          "reasons": {
            "items": {
              "$ref": "#/components/schemas/ConfidenceReason"
            },
            "type": "array",
            "title": "Reasons"
          },
          "links_analyzed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Links Analyzed"
          }
        },
        "type": "object",
        "required": [
          "score"
        ],
        "title": "ConfidenceOut"
      },
      "ConfidenceReason": {
        "properties": {
          "signal": {
            "type": "string",
            "title": "Signal"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "points": {
            "type": "number",
            "title": "Points"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "signal",
          "label",
          "points"
        ],
        "title": "ConfidenceReason"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IngestIssue": {
        "properties": {
          "index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Index"
          },
          "scope": {
            "type": "string",
            "title": "Scope"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "unknown"
          }
        },
        "type": "object",
        "required": [
          "scope",
          "message"
        ],
        "title": "IngestIssue"
      },
...
      "RiskFactor": {
        "properties": {
          "factor": {
            "type": "string",
            "title": "Factor"
          },
          "points": {
            "type": "number",
            "title": "Points"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "factor",
          "points",
          "detail"
        ],
        "title": "RiskFactor"
      },
      "RiskOut": {
        "properties": {
          "score": {
            "type": "number",
            "title": "Score"
          },
          "level": {
            "type": "string",
            "title": "Level"
          },
          "factors": {
            "items": {
              "$ref": "#/components/schemas/RiskFactor"
            },
            "type": "array",
            "title": "Factors"
          }
        },
        "type": "object",
        "required": [
          "score",
          "level"
        ],
        "title": "RiskOut"
      },
...
      "ScenarioGenerateItem": {
        "properties": {
          "scenario_id": {
            "type": "string",
            "title": "Scenario Id"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          },
          "is_benign": {
            "type": "boolean",
            "title": "Is Benign",
            "default": false
          },
          "events_emitted": {
            "type": "integer",
            "title": "Events Emitted",
            "default": 0
          },
          "events_received": {
            "type": "integer",
            "title": "Events Received",
            "default": 0
          },
          "events_stored": {
            "type": "integer",
            "title": "Events Stored",
            "default": 0
          },
          "duplicates": {
            "type": "integer",
            "title": "Duplicates",
            "default": 0
          },
          "skipped": {
            "type": "integer",
            "title": "Skipped",
            "default": 0
          },
          "detections": {
            "type": "integer",
            "title": "Detections",
            "default": 0
          },
          "chains": {
            "type": "integer",
            "title": "Chains",
            "default": 0
          },
          "attack_chains": {
            "type": "integer",
            "title": "Attack Chains",
            "default": 0
          },
          "chain_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Chain Ids"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/IngestIssue"
            },
            "type": "array",
            "title": "Issues"
          },
          "ground_truth": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ground Truth"
          }
        },
        "type": "object",
        "required": [
          "scenario_id"
        ],
        "title": "ScenarioGenerateItem"
      },
      "ScenarioGenerateOut": {
        "properties": {
          "scenarios": {
            "items": {
              "$ref": "#/components/schemas/ScenarioGenerateItem"
            },
            "type": "array",
            "title": "Scenarios"
          },
          "events_stored": {
            "type": "integer",
            "title": "Events Stored"
          },
          "events_replaced": {
            "type": "integer",
            "title": "Events Replaced",
            "default": 0
          },
          "detections": {
            "type": "integer",
            "title": "Detections"
          },
          "chains": {
            "type": "integer",
            "title": "Chains"
          },
          "attack_chains": {
            "type": "integer",
            "title": "Attack Chains"
          },
          "reconstruction": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reconstruction"
          },
          "duration_seconds": {
            "type": "number",
            "title": "Duration Seconds"
          }
        },
        "type": "object",
        "required": [
          "scenarios",
          "events_stored",
          "detections",
          "chains",
          "attack_chains",
          "duration_seconds"
        ],
        "title": "ScenarioGenerateOut"
      },
      "ScenarioGenerateRequest": {
        "properties": {
          "scenario_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Ids",
            "description": "Subset of scenario ids; all scenarios when omitted"
          },
          "seed": {
            "type": "integer",
            "title": "Seed",
            "default": 42
          },
          "ingest": {
            "type": "boolean",
            "title": "Ingest",
            "default": true
          },
          "reconstruct": {
            "type": "boolean",
            "title": "Reconstruct",
            "default": true
          },
          "include_ground_truth": {
            "type": "boolean",
            "title": "Include Ground Truth",
            "default": false
          },
          "replace_duplicates": {
            "type": "boolean",
            "title": "Replace Duplicates",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ScenarioGenerateRequest"
      },
      "ScenarioResetOut": {
        "properties": {
          "cleared": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Cleared"
          },
          "remaining": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Remaining"
          },
          "reconstruction": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reconstruction"
          }
        },
        "type": "object",
        "required": [
          "cleared",
          "remaining"
        ],
        "title": "ScenarioResetOut"
      },
      "ScenarioResetRequest": {
        "properties": {
          "confirm": {
            "type": "boolean",
            "title": "Confirm",
            "default": true
          },
          "reconstruct": {
            "type": "boolean",
            "title": "Reconstruct",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ScenarioResetRequest"
      },
...
      "TechniqueRef": {
        "properties": {
          "technique_id": {
            "type": "string",
            "title": "Technique Id"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          },
          "tactic": {
            "type": "string",
            "title": "Tactic",
            "default": ""
          },
          "observed": {
            "type": "boolean",
            "title": "Observed",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "technique_id"
        ],
        "title": "TechniqueRef"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}